LayoutTests/http/tests/filesystem/workers/resolve-url-sync-expected.txt

[Worker] Tests using resolveLocalFileSystemSyncURL to obtain an Entry from a URL

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: script-tests/resolve-url-sync.js
[Worker] * Resolving a generated URL.
PASS [Worker] expectedPath is actualPath
PASS [Worker] isFile is true
[Worker] * Resolving test file by hand
PASS [Worker] expectedPath is actualPath
PASS [Worker] isFile is true
[Worker] * Resolving a completely bogus URL.
PASS [Worker] Caught exception
[Worker] * Resolving a URL with the wrong protocol
PASS [Worker] Caught exception
[Worker] * Resolving a URL with no slash between type and file
PASS [Worker] Caught exception
[Worker] * Resolving a URL with no slash between protocol and type
PASS [Worker] Caught exception
[Worker] * Resolve a path using backslashes
PASS [Worker] expectedPath is actualPath
PASS [Worker] isFile is true
[Worker] * Resolve a directory
PASS [Worker] expectedPath is actualPath
PASS [Worker] isDirectory is true
[Worker] * Resolve a path using a trailing slash
PASS [Worker] expectedPath is actualPath
PASS [Worker] isDirectory is true
[Worker] * Resolving a persistent URL.
PASS [Worker] expectedPath is actualPath
PASS [Worker] isDirectory is true
[Worker] Finished running tests.
PASS successfullyParsed is true

TEST COMPLETE
